pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "algo-trading"
version = "0.1.0"
description = "Algorithmic trading platform with microstructure bot"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # WebSockets
    "websockets>=12.0",
    "certifi>=2024.0.0",
    # Data validation & config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    # Performance
    "uvloop>=0.21.0",
    # Logging
    "structlog>=24.4.0",
    # HTTP client
    "httpx>=0.28.0",
    # AI
    "anthropic>=0.40.0",
    # Broker integrations
    "ccxt>=4.4.0",
    "alpaca-py>=0.33.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[project.scripts]
micro-runner = "workers.micro_runner:main"

[tool.hatch.build.targets.wheel]
packages = ["backend", "workers"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
